Move configs to config package
