+++ acceptorwrapper/rate_limiting_wrapper.go
@@ -21,8 +21,6 @@
 package acceptorwrapper
 
 import (
-	"time"
-
 	"github.com/topfreegames/pitaya/v2"
 	"github.com/topfreegames/pitaya/v2/acceptor"
 	"github.com/topfreegames/pitaya/v2/config"
@@ -34,33 +32,8 @@ type RateLimitingWrapper struct {
 	BaseWrapper
 }
 
-// RateLimitingConfig rate limits config
-type RateLimitingConfig struct {
-	Limit        int
-	Interval     time.Duration
-	ForceDisable bool
-}
-
-// NewDefaultRateLimitingConfig rate limits default config
-func NewDefaultRateLimitingConfig() RateLimitingConfig {
-	return RateLimitingConfig{
-		Limit:        20,
-		Interval:     time.Duration(time.Second),
-		ForceDisable: false,
-	}
-}
-
-// NewRateLimitingConfig reads from config to build rate limiting configuration
-func NewRateLimitingConfig(config *config.Config) RateLimitingConfig {
-	return RateLimitingConfig{
-		Limit:        config.GetInt("pitaya.conn.ratelimiting.limit"),
-		Interval:     config.GetDuration("pitaya.conn.ratelimiting.interval"),
-		ForceDisable: config.GetBool("pitaya.conn.ratelimiting.forcedisable"),
-	}
-}
-
 // NewRateLimitingWrapper returns an instance of *RateLimitingWrapper
-func NewRateLimitingWrapper(app pitaya.Pitaya, c RateLimitingConfig) *RateLimitingWrapper {
+func NewRateLimitingWrapper(app pitaya.Pitaya, c config.RateLimitingConfig) *RateLimitingWrapper {
 	r := &RateLimitingWrapper{}
 
 	r.BaseWrapper = NewBaseWrapper(func(conn acceptor.PlayerConn) acceptor.PlayerConn {
