+++ app_test.go
@@ -37,6 +37,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/topfreegames/pitaya/v2/acceptor"
 	"github.com/topfreegames/pitaya/v2/cluster"
+	"github.com/topfreegames/pitaya/v2/config"
 	"github.com/topfreegames/pitaya/v2/conn/message"
 	"github.com/topfreegames/pitaya/v2/constants"
 	e "github.com/topfreegames/pitaya/v2/errors"
@@ -70,8 +71,8 @@ func TestMain(m *testing.M) {
 func TestNewApp(t *testing.T) {
 	for _, table := range tables {
 		t.Run(table.serverType, func(t *testing.T) {
-			config := NewDefaultBuilderConfig()
-			app := NewDefaultApp(table.isFrontend, table.serverType, table.serverMode, table.serverMetadata, config).(*App)
+			builderConfig := config.NewDefaultBuilderConfig()
+			app := NewDefaultApp(table.isFrontend, table.serverType, table.serverMode, table.serverMetadata, builderConfig).(*App)
 			assert.Equal(t, table.isFrontend, app.server.Frontend)
 			assert.Equal(t, table.serverType, app.server.Type)
 			assert.Equal(t, table.serverMode, app.serverMode)
@@ -84,8 +85,8 @@ func TestAddAcceptor(t *testing.T) {
 	acc := acceptor.NewTCPAcceptor("0.0.0.0:0")
 	for _, table := range tables {
 		t.Run(table.serverType, func(t *testing.T) {
-			config := NewDefaultBuilderConfig()
-			builder := NewDefaultBuilder(table.isFrontend, table.serverType, table.serverMode, table.serverMetadata, config)
+			builderConfig := config.NewDefaultBuilderConfig()
+			builder := NewDefaultBuilder(table.isFrontend, table.serverType, table.serverMode, table.serverMetadata, builderConfig)
 			builder.AddAcceptor(acc)
 			app := builder.Build().(*App)
 			if table.isFrontend {
@@ -98,8 +99,8 @@ func TestAddAcceptor(t *testing.T) {
 }
 
 func TestSetDebug(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	app.SetDebug(true)
 	assert.Equal(t, true, app.debug)
 	app.SetDebug(false)
@@ -113,33 +114,33 @@ func TestSetLogger(t *testing.T) {
 }
 
 func TestGetDieChan(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	assert.Equal(t, app.dieChan, app.GetDieChan())
 }
 
 func TestGetSever(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	assert.Equal(t, app.server, app.GetServer())
 }
 
 func TestGetMetricsReporters(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	assert.Equal(t, app.metricsReporters, app.GetMetricsReporters())
 }
 func TestGetServerByID(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig)
 	s, err := app.GetServerByID("id")
 	assert.Nil(t, s)
 	assert.EqualError(t, constants.ErrNoServerWithID, err.Error())
 }
 
 func TestGetServersByType(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig)
 	s, err := app.GetServersByType("id")
 	assert.Nil(t, s)
 	assert.EqualError(t, constants.ErrNoServersAvailableOfType, err.Error())
@@ -147,22 +148,22 @@ func TestGetServersByType(t *testing.T) {
 
 func TestSetHeartbeatInterval(t *testing.T) {
 	inter := 35 * time.Millisecond
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	app.SetHeartbeatTime(inter)
 	assert.Equal(t, inter, app.heartbeat)
 }
 
 func TestInitSysRemotes(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	app.initSysRemotes()
 	assert.NotNil(t, app.remoteComp[0])
 }
 
 func TestSetDictionary(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 
 	dict := map[string]uint16{"someroute": 12}
 	err := app.SetDictionary(dict)
@@ -175,8 +176,8 @@ func TestSetDictionary(t *testing.T) {
 }
 
 func TestAddRoute(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	app.router = nil
 	err := app.AddRoute("somesv", func(ctx context.Context, route *route.Route, payload []byte, servers map[string]*cluster.Server) (*cluster.Server, error) {
 		return nil, nil
@@ -197,8 +198,8 @@ func TestAddRoute(t *testing.T) {
 }
 
 func TestShutdown(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	go func() {
 		app.Shutdown()
 	}()
@@ -215,10 +216,10 @@ func TestConfigureDefaultMetricsReporter(t *testing.T) {
 
 	for _, table := range tables {
 		t.Run(fmt.Sprintf("%t", table.enabled), func(t *testing.T) {
-			config := NewDefaultBuilderConfig()
-			config.IsPrometheusEnabled = table.enabled
-			config.IsStatsdEnabled = table.enabled
-			app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+			builderConfig := config.NewDefaultBuilderConfig()
+			builderConfig.IsPrometheusEnabled = table.enabled
+			builderConfig.IsStatsdEnabled = table.enabled
+			app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 			// if statsd is enabled there are 2 metricsReporters, prometheus and statsd
 			assert.Equal(t, table.enabled, len(app.metricsReporters) == 2)
 		})
@@ -226,11 +227,11 @@ func TestConfigureDefaultMetricsReporter(t *testing.T) {
 }
 
 func TestDefaultSD(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	assert.NotNil(t, app.serviceDiscovery)
 
-	etcdSD, err := cluster.NewEtcdServiceDiscovery(cluster.NewDefaultEtcdServiceDiscoveryConfig(), app.server, app.dieChan)
+	etcdSD, err := cluster.NewEtcdServiceDiscovery(config.NewDefaultEtcdServiceDiscoveryConfig(), app.server, app.dieChan)
 	assert.NoError(t, err)
 	typeOfetcdSD := reflect.TypeOf(etcdSD)
 
@@ -240,13 +241,13 @@ func TestDefaultSD(t *testing.T) {
 func TestDefaultRPCServer(t *testing.T) {
 	ctrl := gomock.NewController(t)
 
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	assert.NotNil(t, app.rpcServer)
 
 	sessionPool := mocks.NewMockSessionPool(ctrl)
 
-	natsRPCServer, err := cluster.NewNatsRPCServer(cluster.NewDefaultNatsRPCServerConfig(), app.server, nil, app.dieChan, sessionPool)
+	natsRPCServer, err := cluster.NewNatsRPCServer(config.NewDefaultNatsRPCServerConfig(), app.server, nil, app.dieChan, sessionPool)
 	assert.NoError(t, err)
 	typeOfNatsRPCServer := reflect.TypeOf(natsRPCServer)
 
@@ -254,11 +255,11 @@ func TestDefaultRPCServer(t *testing.T) {
 }
 
 func TestDefaultRPCClient(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	assert.NotNil(t, app.rpcClient)
 
-	natsRPCClient, err := cluster.NewNatsRPCClient(cluster.NewDefaultNatsRPCClientConfig(), app.server, nil, app.dieChan)
+	natsRPCClient, err := cluster.NewNatsRPCClient(config.NewDefaultNatsRPCClientConfig(), app.server, nil, app.dieChan)
 	assert.NoError(t, err)
 	typeOfNatsRPCClient := reflect.TypeOf(natsRPCClient)
 
@@ -266,10 +267,10 @@ func TestDefaultRPCClient(t *testing.T) {
 }
 
 func TestStartAndListenStandalone(t *testing.T) {
-	config := NewDefaultBuilderConfig()
+	builderConfig := config.NewDefaultBuilderConfig()
 
 	acc := acceptor.NewTCPAcceptor("0.0.0.0:0")
-	builder := NewDefaultBuilder(true, "testtype", Standalone, map[string]string{}, config)
+	builder := NewDefaultBuilder(true, "testtype", Standalone, map[string]string{}, builderConfig)
 	builder.AddAcceptor(acc)
 	app := builder.Build().(*App)
 
@@ -298,24 +299,24 @@ func TestStartAndListenCluster(t *testing.T) {
 	ns := helpers.GetTestNatsServer(t)
 	nsAddr := ns.Addr().String()
 
-	builder := NewDefaultBuilder(true, "testtype", Cluster, map[string]string{}, NewDefaultBuilderConfig())
+	builder := NewDefaultBuilder(true, "testtype", Cluster, map[string]string{}, config.NewDefaultBuilderConfig())
 
 	var err error
-	natsClientConfig := cluster.NewDefaultNatsRPCClientConfig()
+	natsClientConfig := config.NewDefaultNatsRPCClientConfig()
 	natsClientConfig.Connect = fmt.Sprintf("nats://%s", nsAddr)
 	builder.RPCClient, err = cluster.NewNatsRPCClient(natsClientConfig, builder.Server, builder.MetricsReporters, builder.DieChan)
 	if err != nil {
 		panic(err.Error())
 	}
 
-	natsServerConfig := cluster.NewDefaultNatsRPCServerConfig()
+	natsServerConfig := config.NewDefaultNatsRPCServerConfig()
 	natsServerConfig.Connect = fmt.Sprintf("nats://%s", nsAddr)
 	builder.RPCServer, err = cluster.NewNatsRPCServer(natsServerConfig, builder.Server, builder.MetricsReporters, builder.DieChan, builder.SessionPool)
 	if err != nil {
 		panic(err.Error())
 	}
 
-	etcdSD, err := cluster.NewEtcdServiceDiscovery(cluster.NewDefaultEtcdServiceDiscoveryConfig(), builder.Server, builder.DieChan, cli)
+	etcdSD, err := cluster.NewEtcdServiceDiscovery(config.NewDefaultEtcdServiceDiscoveryConfig(), builder.Server, builder.DieChan, cli)
 	builder.ServiceDiscovery = etcdSD
 	assert.NoError(t, err)
 	acc := acceptor.NewTCPAcceptor("0.0.0.0:0")
@@ -374,7 +375,7 @@ func TestGetSessionFromCtx(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	ss := mocks.NewMockSession(ctrl)
 
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, NewDefaultBuilderConfig())
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config.NewDefaultBuilderConfig())
 	ctx := context.WithValue(context.Background(), constants.SessionCtxKey, ss)
 	s := app.GetSessionFromCtx(ctx)
 	assert.Equal(t, ss, s)
@@ -423,8 +424,8 @@ func TestDescriptor(t *testing.T) {
 }
 
 func TestDocumentation(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	app.startupComponents()
 	doc, err := app.Documentation(false)
 	assert.NoError(t, err)
@@ -483,8 +484,8 @@ func TestDocumentation(t *testing.T) {
 }
 
 func TestDocumentationTrue(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 	app.startupComponents()
 	doc, err := app.Documentation(true)
 	assert.NoError(t, err)
@@ -579,8 +580,8 @@ func TestAddGRPCInfoToMetadata(t *testing.T) {
 }
 
 func TestStartWorker(t *testing.T) {
-	config := NewDefaultBuilderConfig()
-	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
+	builderConfig := config.NewDefaultBuilderConfig()
+	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig).(*App)
 
 	app.StartWorker()
 	assert.True(t, app.worker.Started())
@@ -588,8 +589,8 @@ func TestStartWorker(t *testing.T) {
 
 func TestRegisterRPCJob(t *testing.T) {
 	t.Run("register_once", func(t *testing.T) {
-		config := NewDefaultBuilderConfig()
-		app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config)
+		builderConfig := config.NewDefaultBuilderConfig()
+		app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig)
 		app.StartWorker()
 
 		err := app.RegisterRPCJob(nil)
@@ -597,8 +598,8 @@ func TestRegisterRPCJob(t *testing.T) {
 	})
 
 	t.Run("register_twice", func(t *testing.T) {
-		config := NewDefaultBuilderConfig()
-		app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config)
+		builderConfig := config.NewDefaultBuilderConfig()
+		app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, builderConfig)
 		app.StartWorker()
 
 		err := app.RegisterRPCJob(nil)
