+++ cluster/grpc_rpc_server_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/golang/mock/gomock"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/topfreegames/pitaya/v2/config"
 	"github.com/topfreegames/pitaya/v2/helpers"
 	"github.com/topfreegames/pitaya/v2/metrics"
 	protosmocks "github.com/topfreegames/pitaya/v2/protos/mocks"
@@ -16,14 +17,14 @@ import (
 func TestNewGRPCServer(t *testing.T) {
 	t.Parallel()
 	sv := getServer()
-	gs, err := NewGRPCServer(NewDefaultGRPCServerConfig(), sv, []metrics.Reporter{})
+	gs, err := NewGRPCServer(config.NewDefaultGRPCServerConfig(), sv, []metrics.Reporter{})
 	assert.NoError(t, err)
 	assert.NotNil(t, gs)
 }
 
 func TestGRPCServerInit(t *testing.T) {
 	t.Parallel()
-	c := NewDefaultGRPCServerConfig()
+	c := config.NewDefaultGRPCServerConfig()
 	c.Port = helpers.GetFreePort(t)
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
