+++ component_test.go
@@ -25,6 +25,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/topfreegames/pitaya/v2/component"
+	"github.com/topfreegames/pitaya/v2/config"
 )
 
 type MyComp struct {
@@ -41,7 +42,7 @@ func (m *MyComp) Shutdown() {
 }
 
 func TestRegister(t *testing.T) {
-	config := NewDefaultBuilderConfig()
+	config := config.NewDefaultBuilderConfig()
 	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
 	b := &component.Base{}
 	app.Register(b)
@@ -50,7 +51,7 @@ func TestRegister(t *testing.T) {
 }
 
 func TestRegisterRemote(t *testing.T) {
-	config := NewDefaultBuilderConfig()
+	config := config.NewDefaultBuilderConfig()
 	app := NewDefaultApp(true, "testtype", Cluster, map[string]string{}, config).(*App)
 	before := app.remoteComp
 	b := &component.Base{}
@@ -60,7 +61,7 @@ func TestRegisterRemote(t *testing.T) {
 }
 
 func TestStartupComponents(t *testing.T) {
-	app := NewDefaultApp(true, "testtype", Standalone, map[string]string{}, NewDefaultBuilderConfig()).(*App)
+	app := NewDefaultApp(true, "testtype", Standalone, map[string]string{}, config.NewDefaultBuilderConfig()).(*App)
 
 	app.Register(&MyComp{})
 	app.RegisterRemote(&MyComp{})
@@ -69,7 +70,7 @@ func TestStartupComponents(t *testing.T) {
 }
 
 func TestShutdownComponents(t *testing.T) {
-	app := NewDefaultApp(true, "testtype", Standalone, map[string]string{}, NewDefaultBuilderConfig()).(*App)
+	app := NewDefaultApp(true, "testtype", Standalone, map[string]string{}, config.NewDefaultBuilderConfig()).(*App)
 
 	app.Register(&MyComp{})
 	app.RegisterRemote(&MyComp{})
