+++ config/config.go
@@ -1,173 +1,504 @@
-// Copyright (c) TFG Co. All Rights Reserved.
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
 package config
 
 import (
-	"strings"
+	"strconv"
 	"time"
 
-	"github.com/spf13/viper"
+	"github.com/topfreegames/pitaya/v2/metrics/models"
 )
 
-// Config is a wrapper around a viper config
-type Config struct {
-	config *viper.Viper
+// PitayaConfig provides configuration for a pitaya app
+type PitayaConfig struct {
+	HearbeatInterval           time.Duration
+	MessageCompression         bool
+	BufferAgentMessages        int
+	BufferHandlerLocalProcess  int
+	BufferHandlerRemoteProcess int
+	ConcurrencyHandlerDispatch int
+	SessionUnique              bool
+	MetricsPeriod              time.Duration
 }
 
-// NewConfig creates a new config with a given viper config if given
-func NewConfig(cfgs ...*viper.Viper) *Config {
-	var cfg *viper.Viper
-	if len(cfgs) > 0 {
-		cfg = cfgs[0]
-	} else {
-		cfg = viper.New()
-	}
-
-	cfg.SetEnvKeyReplacer(strings.NewReplacer(".", "_"))
-	cfg.AutomaticEnv()
-	c := &Config{config: cfg}
-	c.fillDefaultValues()
-	return c
-}
-
-func (c *Config) fillDefaultValues() {
-	defaultsMap := map[string]interface{}{
-		"pitaya.buffer.agent.messages": 100,
-		// the max buffer size that nats will accept, if this buffer overflows, messages will begin to be dropped
-		"pitaya.buffer.cluster.rpc.server.nats.messages":        75,
-		"pitaya.buffer.cluster.rpc.server.nats.push":            100,
-		"pitaya.buffer.handler.localprocess":                    20,
-		"pitaya.buffer.handler.remoteprocess":                   20,
-		"pitaya.cluster.info.region":                            "",
-		"pitaya.cluster.rpc.client.grpc.dialtimeout":            "5s",
-		"pitaya.cluster.rpc.client.grpc.requesttimeout":         "5s",
-		"pitaya.cluster.rpc.client.grpc.lazyconnection":         false,
-		"pitaya.cluster.rpc.client.nats.connect":                "nats://localhost:4222",
-		"pitaya.cluster.rpc.client.nats.connectiontimeout":      "2s",
-		"pitaya.cluster.rpc.client.nats.maxreconnectionretries": 15,
-		"pitaya.cluster.rpc.client.nats.requesttimeout":         "5s",
-		"pitaya.cluster.rpc.server.grpc.externalport":           3434,
-		"pitaya.cluster.rpc.server.grpc.port":                   3434,
-		"pitaya.cluster.rpc.server.nats.connect":                "nats://localhost:4222",
-		"pitaya.cluster.rpc.server.nats.connectiontimeout":      "2s",
-		"pitaya.cluster.rpc.server.nats.maxreconnectionretries": 15,
-		"pitaya.cluster.sd.etcd.dialtimeout":                    "5s",
-		"pitaya.cluster.sd.etcd.endpoints":                      "localhost:2379",
-		"pitaya.cluster.sd.etcd.grantlease.maxretries":          15,
-		"pitaya.cluster.sd.etcd.grantlease.retryinterval":       "5s",
-		"pitaya.cluster.sd.etcd.grantlease.timeout":             "60s",
-		"pitaya.cluster.sd.etcd.heartbeat.log":                  false,
-		"pitaya.cluster.sd.etcd.heartbeat.ttl":                  "60s",
-		"pitaya.cluster.sd.etcd.prefix":                         "pitaya/",
-		"pitaya.cluster.sd.etcd.revoke.timeout":                 "5s",
-		"pitaya.cluster.sd.etcd.syncservers.interval":           "120s",
-		"pitaya.cluster.sd.etcd.shutdown.delay":                 "10ms",
-		"pitaya.cluster.sd.etcd.servertypeblacklist":            nil,
-		"pitaya.cluster.sd.etcd.syncserversparallelism":         10,
-		// the sum of this config among all the frontend servers should always be less than
-		// the sum of pitaya.buffer.cluster.rpc.server.nats.messages, for covering the worst case scenario
-		// a single backend server should have the config pitaya.buffer.cluster.rpc.server.nats.messages bigger
-		// than the sum of the config pitaya.concurrency.handler.dispatch among all frontend servers
-		"pitaya.concurrency.handler.dispatch":              25,
-		"pitaya.concurrency.remote.service":                30,
-		"pitaya.defaultpipelines.structvalidation.enabled": false,
-		"pitaya.groups.etcd.dialtimeout":                   "5s",
-		"pitaya.groups.etcd.endpoints":                     "localhost:2379",
-		"pitaya.groups.etcd.prefix":                        "pitaya/",
-		"pitaya.groups.etcd.transactiontimeout":            "5s",
-		"pitaya.groups.memory.tickduration":                "30s",
-		"pitaya.handler.messages.compression":              true,
-		"pitaya.heartbeat.interval":                        "30s",
-		"pitaya.metrics.additionalTags":                    map[string]string{},
-		"pitaya.metrics.constTags":                         map[string]string{},
-		"pitaya.metrics.custom":                            map[string]interface{}{},
-		"pitaya.metrics.periodicMetrics.period":            "15s",
-		"pitaya.metrics.prometheus.enabled":                false,
-		"pitaya.metrics.prometheus.port":                   9090,
-		"pitaya.metrics.statsd.enabled":                    false,
-		"pitaya.metrics.statsd.host":                       "localhost:9125",
-		"pitaya.metrics.statsd.prefix":                     "pitaya.",
-		"pitaya.metrics.statsd.rate":                       1,
-		"pitaya.modules.bindingstorage.etcd.dialtimeout":   "5s",
-		"pitaya.modules.bindingstorage.etcd.endpoints":     "localhost:2379",
-		"pitaya.modules.bindingstorage.etcd.leasettl":      "1h",
-		"pitaya.modules.bindingstorage.etcd.prefix":        "pitaya/",
-		"pitaya.conn.ratelimiting.limit":                   20,
-		"pitaya.conn.ratelimiting.interval":                "1s",
-		"pitaya.conn.ratelimiting.forcedisable":            false,
-		"pitaya.session.unique":                            true,
-		"pitaya.worker.concurrency":                        1,
-		"pitaya.worker.redis.pool":                         "10",
-		"pitaya.worker.redis.url":                          "localhost:6379",
-		"pitaya.worker.retry.enabled":                      true,
-		"pitaya.worker.retry.exponential":                  2,
-		"pitaya.worker.retry.max":                          5,
-		"pitaya.worker.retry.maxDelay":                     10,
-		"pitaya.worker.retry.maxRandom":                    10,
-		"pitaya.worker.retry.minDelay":                     0,
-	}
-
-	for param := range defaultsMap {
-		if c.config.Get(param) == nil {
-			c.config.SetDefault(param, defaultsMap[param])
-		}
+// NewDefaultPitayaConfig provides default configuration for Pitaya App
+func NewDefaultPitayaConfig() PitayaConfig {
+	return PitayaConfig{
+		HearbeatInterval:           time.Duration(30 * time.Second),
+		MessageCompression:         true,
+		BufferAgentMessages:        100,
+		BufferHandlerLocalProcess:  20,
+		BufferHandlerRemoteProcess: 20,
+		ConcurrencyHandlerDispatch: 25,
+		SessionUnique:              true,
+		MetricsPeriod:              time.Duration(15 * time.Second),
 	}
 }
 
-// GetDuration returns a duration from the inner config
-func (c *Config) GetDuration(s string) time.Duration {
-	return c.config.GetDuration(s)
+// NewPitayaConfig returns a config instance with values extracted from default config paths
+func NewPitayaConfig(config *Config) PitayaConfig {
+	return PitayaConfig{
+		HearbeatInterval:           config.GetDuration("pitaya.heartbeat.interval"),
+		MessageCompression:         config.GetBool("pitaya.handler.messages.compression"),
+		BufferAgentMessages:        config.GetInt("pitaya.buffer.agent.messages"),
+		BufferHandlerLocalProcess:  config.GetInt("pitaya.buffer.handler.localprocess"),
+		BufferHandlerRemoteProcess: config.GetInt("pitaya.buffer.handler.remoteprocess"),
+		ConcurrencyHandlerDispatch: config.GetInt("pitaya.concurrency.handler.dispatch"),
+		SessionUnique:              config.GetBool("pitaya.session.unique"),
+		MetricsPeriod:              config.GetDuration("pitaya.metrics.periodicMetrics.period"),
+	}
 }
 
-// GetString returns a string from the inner config
-func (c *Config) GetString(s string) string {
-	return c.config.GetString(s)
+// BuilderConfig provides configuration for Builder
+type BuilderConfig struct {
+	PitayaConfig             PitayaConfig
+	IsPrometheusEnabled      bool
+	IsStatsdEnabled          bool
+	IsDefaultPipelineEnabled bool
 }
 
-// GetInt returns an int from the inner config
-func (c *Config) GetInt(s string) int {
-	return c.config.GetInt(s)
+// NewDefaultBuilderConfig provides default builder configuration
+func NewDefaultBuilderConfig() BuilderConfig {
+	return BuilderConfig{
+		PitayaConfig:             NewDefaultPitayaConfig(),
+		IsPrometheusEnabled:      false,
+		IsStatsdEnabled:          false,
+		IsDefaultPipelineEnabled: false,
+	}
 }
 
-// GetBool returns an boolean from the inner config
-func (c *Config) GetBool(s string) bool {
-	return c.config.GetBool(s)
+// NewBuilderConfig reads from config to build builder configuration
+func NewBuilderConfig(config *Config) BuilderConfig {
+	return BuilderConfig{
+		PitayaConfig:             NewPitayaConfig(config),
+		IsPrometheusEnabled:      config.GetBool("pitaya.metrics.prometheus.enabled"),
+		IsStatsdEnabled:          config.GetBool("pitaya.metrics.statsd.enabled"),
+		IsDefaultPipelineEnabled: config.GetBool("pitaya.defaultpipelines.structvalidation.enabled"),
+	}
 }
 
-// GetStringSlice returns a string slice from the inner config
-func (c *Config) GetStringSlice(s string) []string {
-	return c.config.GetStringSlice(s)
+// GRPCClientConfig rpc client config struct
+type GRPCClientConfig struct {
+	DialTimeout    time.Duration
+	LazyConnection bool
+	RequestTimeout time.Duration
 }
 
-// Get returns an interface from the inner config
-func (c *Config) Get(s string) interface{} {
-	return c.config.Get(s)
+// NewDefaultGRPCClientConfig rpc client default config struct
+func NewDefaultGRPCClientConfig() GRPCClientConfig {
+	return GRPCClientConfig{
+		DialTimeout:    time.Duration(5 * time.Second),
+		LazyConnection: false,
+		RequestTimeout: time.Duration(5 * time.Second),
+	}
 }
 
-// GetStringMapString returns a string map string from the inner config
-func (c *Config) GetStringMapString(s string) map[string]string {
-	return c.config.GetStringMapString(s)
+// NewGRPCClientConfig reads from config to build GRPCCLientConfig
+func NewGRPCClientConfig(config *Config) GRPCClientConfig {
+	return GRPCClientConfig{
+		DialTimeout:    config.GetDuration("pitaya.cluster.rpc.client.grpc.dialtimeout"),
+		LazyConnection: config.GetBool("pitaya.cluster.rpc.client.grpc.lazyconnection"),
+		RequestTimeout: config.GetDuration("pitaya.cluster.rpc.client.grpc.requesttimeout"),
+	}
 }
 
-// UnmarshalKey unmarshals key into v
-func (c *Config) UnmarshalKey(s string, v interface{}) error {
-	return c.config.UnmarshalKey(s, v)
+// GRPCServerConfig provides configuration for GRPCServer
+type GRPCServerConfig struct {
+	Port int
+}
+
+// NewDefaultGRPCServerConfig returns a default GRPCServerConfig
+func NewDefaultGRPCServerConfig() GRPCServerConfig {
+	return GRPCServerConfig{
+		Port: 3434,
+	}
+}
+
+// NewGRPCServerConfig reads from config to build GRPCServerConfig
+func NewGRPCServerConfig(config *Config) GRPCServerConfig {
+	return GRPCServerConfig{
+		Port: config.GetInt("pitaya.cluster.rpc.server.grpc.port"),
+	}
+}
+
+// NatsRPCClientConfig provides nats client configuration
+type NatsRPCClientConfig struct {
+	Connect                string
+	MaxReconnectionRetries int
+	RequestTimeout         time.Duration
+	ConnectionTimeout      time.Duration
+}
+
+// NewDefaultNatsRPCClientConfig provides default nats client configuration
+func NewDefaultNatsRPCClientConfig() NatsRPCClientConfig {
+	return NatsRPCClientConfig{
+		Connect:                "nats://localhost:4222",
+		MaxReconnectionRetries: 15,
+		RequestTimeout:         time.Duration(5 * time.Second),
+		ConnectionTimeout:      time.Duration(2 * time.Second),
+	}
+}
+
+// NewNatsRPCClientConfig reads from config to build nats client configuration
+func NewNatsRPCClientConfig(config *Config) NatsRPCClientConfig {
+	return NatsRPCClientConfig{
+		Connect:                config.GetString("pitaya.cluster.rpc.client.nats.connect"),
+		MaxReconnectionRetries: config.GetInt("pitaya.cluster.rpc.client.nats.maxreconnectionretries"),
+		RequestTimeout:         config.GetDuration("pitaya.cluster.rpc.client.nats.requesttimeout"),
+		ConnectionTimeout:      config.GetDuration("pitaya.cluster.rpc.client.nats.connectiontimeout"),
+	}
+}
+
+// NatsRPCServerConfig provides nats server configuration
+type NatsRPCServerConfig struct {
+	Connect                string
+	MaxReconnectionRetries int
+	Messages               int
+	Push                   int
+	Service                int
+	ConnectionTimeout      time.Duration
+}
+
+// NewDefaultNatsRPCServerConfig provides default nats server configuration
+func NewDefaultNatsRPCServerConfig() NatsRPCServerConfig {
+	return NatsRPCServerConfig{
+		Connect:                "nats://localhost:4222",
+		MaxReconnectionRetries: 15,
+		Messages:               75,
+		Push:                   100,
+		Service:                30,
+		ConnectionTimeout:      time.Duration(2 * time.Second),
+	}
+}
+
+// NewNatsRPCServerConfig reads from config to build nats server configuration
+func NewNatsRPCServerConfig(config *Config) NatsRPCServerConfig {
+	return NatsRPCServerConfig{
+		Connect:                config.GetString("pitaya.cluster.rpc.server.nats.connect"),
+		MaxReconnectionRetries: config.GetInt("pitaya.cluster.rpc.server.nats.maxreconnectionretries"),
+		Messages:               config.GetInt("pitaya.buffer.cluster.rpc.server.nats.messages"),
+		Push:                   config.GetInt("pitaya.buffer.cluster.rpc.server.nats.push"),
+		Service:                config.GetInt("pitaya.concurrency.remote.service"),
+		ConnectionTimeout:      config.GetDuration("pitaya.cluster.rpc.server.nats.connectiontimeout"),
+	}
+}
+
+// ConfigInfoRetrieverConfig provides ConfigInfoRetriever configuration
+type ConfigInfoRetrieverConfig struct {
+	Region string
+}
+
+// NewDefaultConfigInfoRetrieverConfig provides default configuration for ConfigInfoRetriever
+func NewDefaultConfigInfoRetrieverConfig() ConfigInfoRetrieverConfig {
+	return ConfigInfoRetrieverConfig{
+		Region: "",
+	}
+}
+
+// NewConfigInfoRetrieverConfig reads from config to build configuration for ConfigInfoRetriever
+func NewConfigInfoRetrieverConfig(c *Config) ConfigInfoRetrieverConfig {
+	return ConfigInfoRetrieverConfig{
+		Region: c.GetString("pitaya.cluster.info.region"),
+	}
+}
+
+// EtcdServiceDiscoveryConfig Etcd service discovery config
+type EtcdServiceDiscoveryConfig struct {
+	EtcdEndpoints          []string
+	EtcdUser               string
+	EtcdPass               string
+	EtcdDialTimeout        time.Duration
+	EtcdPrefix             string
+	HeartbeatTTL           time.Duration
+	LogHeartbeat           bool
+	SyncServersInterval    time.Duration
+	RevokeTimeout          time.Duration
+	GrantLeaseTimeout      time.Duration
+	GrantLeaseMaxRetries   int
+	GrantLeaseInterval     time.Duration
+	ShutdownDelay          time.Duration
+	ServerTypesBlacklist   []string
+	SyncServersParallelism int
+}
+
+// NewDefaultEtcdServiceDiscoveryConfig Etcd service discovery default config
+func NewDefaultEtcdServiceDiscoveryConfig() EtcdServiceDiscoveryConfig {
+	return EtcdServiceDiscoveryConfig{
+		EtcdEndpoints:          []string{"localhost:2379"},
+		EtcdUser:               "",
+		EtcdPass:               "",
+		EtcdDialTimeout:        time.Duration(5 * time.Second),
+		EtcdPrefix:             "pitaya/",
+		HeartbeatTTL:           time.Duration(60 * time.Second),
+		LogHeartbeat:           false,
+		SyncServersInterval:    time.Duration(120 * time.Second),
+		RevokeTimeout:          time.Duration(5 * time.Second),
+		GrantLeaseTimeout:      time.Duration(60 * time.Second),
+		GrantLeaseMaxRetries:   15,
+		GrantLeaseInterval:     time.Duration(5 * time.Second),
+		ShutdownDelay:          time.Duration(10 * time.Millisecond),
+		ServerTypesBlacklist:   nil,
+		SyncServersParallelism: 10,
+	}
+}
+
+// NewEtcdServiceDiscoveryConfig Etcd service discovery config with default config paths
+func NewEtcdServiceDiscoveryConfig(config *Config) EtcdServiceDiscoveryConfig {
+	return EtcdServiceDiscoveryConfig{
+		EtcdEndpoints:          config.GetStringSlice("pitaya.cluster.sd.etcd.endpoints"),
+		EtcdUser:               config.GetString("pitaya.cluster.sd.etcd.user"),
+		EtcdPass:               config.GetString("pitaya.cluster.sd.etcd.pass"),
+		EtcdDialTimeout:        config.GetDuration("pitaya.cluster.sd.etcd.dialtimeout"),
+		EtcdPrefix:             config.GetString("pitaya.cluster.sd.etcd.prefix"),
+		HeartbeatTTL:           config.GetDuration("pitaya.cluster.sd.etcd.heartbeat.ttl"),
+		LogHeartbeat:           config.GetBool("pitaya.cluster.sd.etcd.heartbeat.log"),
+		SyncServersInterval:    config.GetDuration("pitaya.cluster.sd.etcd.syncservers.interval"),
+		RevokeTimeout:          config.GetDuration("pitaya.cluster.sd.etcd.revoke.timeout"),
+		GrantLeaseTimeout:      config.GetDuration("pitaya.cluster.sd.etcd.grantlease.timeout"),
+		GrantLeaseMaxRetries:   config.GetInt("pitaya.cluster.sd.etcd.grantlease.maxretries"),
+		GrantLeaseInterval:     config.GetDuration("pitaya.cluster.sd.etcd.grantlease.retryinterval"),
+		ShutdownDelay:          config.GetDuration("pitaya.cluster.sd.etcd.shutdown.delay"),
+		ServerTypesBlacklist:   config.GetStringSlice("pitaya.cluster.sd.etcd.servertypeblacklist"),
+		SyncServersParallelism: config.GetInt("pitaya.cluster.sd.etcd.syncserversparallelism"),
+	}
+}
+
+// NewDefaultCustomMetricsSpec returns an empty *CustomMetricsSpec
+func NewDefaultCustomMetricsSpec() models.CustomMetricsSpec {
+	return models.CustomMetricsSpec{
+		Summaries: []*models.Summary{},
+		Gauges:    []*models.Gauge{},
+		Counters:  []*models.Counter{},
+	}
+}
+
+// NewCustomMetricsSpec returns a *CustomMetricsSpec by reading config key (DEPRECATED)
+func NewCustomMetricsSpec(config *Config) models.CustomMetricsSpec {
+	var spec models.CustomMetricsSpec
+
+	err := config.UnmarshalKey("pitaya.metrics.custom", &spec)
+	if err != nil {
+		return NewDefaultCustomMetricsSpec()
+	}
+
+	return spec
+}
+
+// PrometheusConfig provides configuration for PrometheusReporter
+type PrometheusConfig struct {
+	Port             int
+	Game             string
+	AdditionalLabels map[string]string
+	ConstLabels      map[string]string
+}
+
+// NewDefaultPrometheusConfig provides default configuration for PrometheusReporter
+func NewDefaultPrometheusConfig() PrometheusConfig {
+	return PrometheusConfig{
+		Port:             9090,
+		AdditionalLabels: map[string]string{},
+		ConstLabels:      map[string]string{},
+	}
+}
+
+// NewPrometheusConfig reads from config to build configuration for PrometheusReporter
+func NewPrometheusConfig(config *Config) PrometheusConfig {
+	return PrometheusConfig{
+		Port:             config.GetInt("pitaya.metrics.prometheus.port"),
+		Game:             config.GetString("pitaya.game"),
+		AdditionalLabels: config.GetStringMapString("pitaya.metrics.additionalTags"),
+		ConstLabels:      config.GetStringMapString("pitaya.metrics.constTags"),
+	}
+}
+
+// StatsdConfig provides configuration for statsd
+type StatsdConfig struct {
+	Host        string
+	Prefix      string
+	Rate        float64
+	ConstLabels map[string]string
+}
+
+// NewDefaultStatsdConfig provides default configuration for statsd
+func NewDefaultStatsdConfig() StatsdConfig {
+	return StatsdConfig{
+		Host:        "localhost:9125",
+		Prefix:      "pitaya.",
+		Rate:        1,
+		ConstLabels: map[string]string{},
+	}
+}
+
+// NewStatsdConfig reads from config to build configuration for statsd
+func NewStatsdConfig(config *Config) StatsdConfig {
+	rate, err := strconv.ParseFloat(config.GetString("pitaya.metrics.statsd.rate"), 64)
+	if err != nil {
+		panic(err.Error())
+	}
+
+	statsdConfig := StatsdConfig{
+		Host:        config.GetString("pitaya.metrics.statsd.host"),
+		Prefix:      config.GetString("pitaya.metrics.statsd.prefix"),
+		Rate:        rate,
+		ConstLabels: config.GetStringMapString("pitaya.metrics.constTags"),
+	}
+
+	return statsdConfig
+}
+
+// WorkerConfig provides worker configuration
+type WorkerConfig struct {
+	ServerURL   string
+	Pool        string
+	Password    string
+	Namespace   string
+	Concurrency int
+}
+
+// NewDefaultWorkerConfig provides worker default configuration
+func NewDefaultWorkerConfig() WorkerConfig {
+	return WorkerConfig{
+		ServerURL:   "localhost:6379",
+		Pool:        "10",
+		Concurrency: 1,
+	}
+}
+
+// NewWorkerConfig provides worker configuration based on default string paths
+func NewWorkerConfig(config *Config) WorkerConfig {
+	return WorkerConfig{
+		ServerURL:   config.GetString("pitaya.worker.redis.url"),
+		Pool:        config.GetString("pitaya.worker.redis.pool"),
+		Password:    config.GetString("pitaya.worker.redis.password"),
+		Namespace:   config.GetString("pitaya.worker.namespace"),
+		Concurrency: config.GetInt("pitaya.worker.concurrency"),
+	}
+}
+
+// EnqueueOpts has retry options for worker
+type EnqueueOpts struct {
+	RetryEnabled      bool
+	MaxRetries        int
+	ExponentialFactor int
+	MinDelayToRetry   int
+	MaxDelayToRetry   int
+	MaxRandom         int
+}
+
+// NewDefaultEnqueueOpts provides default EnqueueOpts
+func NewDefaultEnqueueOpts() EnqueueOpts {
+	return EnqueueOpts{
+		RetryEnabled:      true,
+		MaxRetries:        2,
+		ExponentialFactor: 5,
+		MinDelayToRetry:   10,
+		MaxDelayToRetry:   10,
+		MaxRandom:         0,
+	}
+}
+
+// NewEnqueueOpts reads from config to build *EnqueueOpts
+func NewEnqueueOpts(config *Config) EnqueueOpts {
+	return EnqueueOpts{
+		RetryEnabled:      config.GetBool("pitaya.worker.retry.enabled"),
+		MaxRetries:        config.GetInt("pitaya.worker.retry.max"),
+		ExponentialFactor: config.GetInt("pitaya.worker.retry.exponential"),
+		MinDelayToRetry:   config.GetInt("pitaya.worker.retry.minDelay"),
+		MaxDelayToRetry:   config.GetInt("pitaya.worker.retry.maxDelay"),
+		MaxRandom:         config.GetInt("pitaya.worker.retry.maxRandom"),
+	}
+}
+
+// MemoryGroupConfig provides configuration for MemoryGroup
+type MemoryGroupConfig struct {
+	TickDuration time.Duration
+}
+
+// NewDefaultMemoryGroupConfig returns a new, default group instance
+func NewDefaultMemoryGroupConfig() MemoryGroupConfig {
+	return MemoryGroupConfig{TickDuration: time.Duration(30 * time.Second)}
+}
+
+// NewMemoryGroupConfig returns a new, default group instance
+func NewMemoryGroupConfig(conf *Config) MemoryGroupConfig {
+	return MemoryGroupConfig{TickDuration: conf.GetDuration("pitaya.groups.memory.tickduration")}
+}
+
+// EtcdGroupServiceConfig provides ETCD configuration
+type EtcdGroupServiceConfig struct {
+	DialTimeout        time.Duration
+	Endpoints          []string
+	Prefix             string
+	TransactionTimeout time.Duration
+}
+
+// NewDefaultEtcdGroupServiceConfig provides default ETCD configuration
+func NewDefaultEtcdGroupServiceConfig() EtcdGroupServiceConfig {
+	return EtcdGroupServiceConfig{
+		DialTimeout:        time.Duration(5 * time.Second),
+		Endpoints:          []string{"localhost:2379"},
+		Prefix:             "pitaya/",
+		TransactionTimeout: time.Duration(5 * time.Second),
+	}
+}
+
+// NewEtcdGroupServiceConfig reads from config to build ETCD configuration
+func NewEtcdGroupServiceConfig(config *Config) EtcdGroupServiceConfig {
+	return EtcdGroupServiceConfig{
+		DialTimeout:        config.GetDuration("pitaya.groups.etcd.dialtimeout"),
+		Endpoints:          config.GetStringSlice("pitaya.groups.etcd.endpoints"),
+		Prefix:             config.GetString("pitaya.groups.etcd.prefix"),
+		TransactionTimeout: config.GetDuration("pitaya.groups.etcd.transactiontimeout"),
+	}
+}
+
+// ETCDBindingConfig provides configuration for ETCDBindingStorage
+type ETCDBindingConfig struct {
+	DialTimeout time.Duration
+	Endpoints   []string
+	Prefix      string
+	LeaseTTL    time.Duration
+}
+
+// NewDefaultETCDBindingConfig provides default configuration for ETCDBindingStorage
+func NewDefaultETCDBindingConfig() ETCDBindingConfig {
+	return ETCDBindingConfig{
+		DialTimeout: time.Duration(5 * time.Second),
+		Endpoints:   []string{"localhost:2379"},
+		Prefix:      "pitaya/",
+		LeaseTTL:    time.Duration(5 * time.Hour),
+	}
+}
+
+// NewETCDBindingConfig reads from config to build ETCDBindingStorage configuration
+func NewETCDBindingConfig(config *Config) ETCDBindingConfig {
+	return ETCDBindingConfig{
+		DialTimeout: config.GetDuration("pitaya.modules.bindingstorage.etcd.dialtimeout"),
+		Endpoints:   config.GetStringSlice("pitaya.modules.bindingstorage.etcd.endpoints"),
+		Prefix:      config.GetString("pitaya.modules.bindingstorage.etcd.prefix"),
+		LeaseTTL:    config.GetDuration("pitaya.modules.bindingstorage.etcd.leasettl"),
+	}
+}
+
+// RateLimitingConfig rate limits config
+type RateLimitingConfig struct {
+	Limit        int
+	Interval     time.Duration
+	ForceDisable bool
+}
+
+// NewDefaultRateLimitingConfig rate limits default config
+func NewDefaultRateLimitingConfig() RateLimitingConfig {
+	return RateLimitingConfig{
+		Limit:        20,
+		Interval:     time.Duration(time.Second),
+		ForceDisable: false,
+	}
+}
+
+// NewRateLimitingConfig reads from config to build rate limiting configuration
+func NewRateLimitingConfig(config *Config) RateLimitingConfig {
+	return RateLimitingConfig{
+		Limit:        config.GetInt("pitaya.conn.ratelimiting.limit"),
+		Interval:     config.GetDuration("pitaya.conn.ratelimiting.interval"),
+		ForceDisable: config.GetBool("pitaya.conn.ratelimiting.forcedisable"),
+	}
 }
