+++ examples/demo/cluster_protobuf/main.go
@@ -9,6 +9,7 @@ import (
 	"github.com/topfreegames/pitaya/v2"
 	"github.com/topfreegames/pitaya/v2/acceptor"
 	"github.com/topfreegames/pitaya/v2/component"
+	"github.com/topfreegames/pitaya/v2/config"
 	"github.com/topfreegames/pitaya/v2/examples/demo/cluster_protobuf/services"
 	"github.com/topfreegames/pitaya/v2/groups"
 	"github.com/topfreegames/pitaya/v2/serialize/protobuf"
@@ -47,13 +48,13 @@ func main() {
 
 	flag.Parse()
 
-	builder := pitaya.NewDefaultBuilder(*isFrontend, *svType, pitaya.Cluster, map[string]string{}, pitaya.NewDefaultBuilderConfig())
+	builder := pitaya.NewDefaultBuilder(*isFrontend, *svType, pitaya.Cluster, map[string]string{}, config.NewDefaultBuilderConfig())
 	if *isFrontend {
 		ws := acceptor.NewWSAcceptor(fmt.Sprintf(":%d", port))
 		builder.AddAcceptor(ws)
 	}
 	builder.Serializer = protobuf.NewSerializer()
-	builder.Groups = groups.NewMemoryGroupService(groups.NewDefaultMemoryGroupConfig())
+	builder.Groups = groups.NewMemoryGroupService(config.NewDefaultMemoryGroupConfig())
 	app := builder.Build()
 
 	defer app.Shutdown()
