+++ groups/etcd_group_service.go
@@ -24,36 +24,8 @@ var (
 type EtcdGroupService struct {
 }
 
-// EtcdGroupServiceConfig provides ETCD configuration
-type EtcdGroupServiceConfig struct {
-	DialTimeout        time.Duration
-	Endpoints          []string
-	Prefix             string
-	TransactionTimeout time.Duration
-}
-
-// NewDefaultEtcdGroupServiceConfig provides default ETCD configuration
-func NewDefaultEtcdGroupServiceConfig() EtcdGroupServiceConfig {
-	return EtcdGroupServiceConfig{
-		DialTimeout:        time.Duration(5 * time.Second),
-		Endpoints:          []string{"localhost:2379"},
-		Prefix:             "pitaya/",
-		TransactionTimeout: time.Duration(5 * time.Second),
-	}
-}
-
-// NewEtcdGroupServiceConfig reads from config to build ETCD configuration
-func NewEtcdGroupServiceConfig(config *config.Config) EtcdGroupServiceConfig {
-	return EtcdGroupServiceConfig{
-		DialTimeout:        config.GetDuration("pitaya.groups.etcd.dialtimeout"),
-		Endpoints:          config.GetStringSlice("pitaya.groups.etcd.endpoints"),
-		Prefix:             config.GetString("pitaya.groups.etcd.prefix"),
-		TransactionTimeout: config.GetDuration("pitaya.groups.etcd.transactiontimeout"),
-	}
-}
-
 // NewEtcdGroupService returns a new group instance
-func NewEtcdGroupService(conf EtcdGroupServiceConfig, clientOrNil *clientv3.Client) (*EtcdGroupService, error) {
+func NewEtcdGroupService(conf config.EtcdGroupServiceConfig, clientOrNil *clientv3.Client) (*EtcdGroupService, error) {
 	err := initClientInstance(conf, clientOrNil)
 	if err != nil {
 		return nil, err
@@ -61,7 +33,7 @@ func NewEtcdGroupService(conf EtcdGroupServiceConfig, clientOrNil *clientv3.Clie
 	return &EtcdGroupService{}, err
 }
 
-func initClientInstance(config EtcdGroupServiceConfig, clientOrNil *clientv3.Client) error {
+func initClientInstance(config config.EtcdGroupServiceConfig, clientOrNil *clientv3.Client) error {
 	var err error
 	etcdOnce.Do(func() {
 		if clientOrNil != nil {
@@ -78,7 +50,7 @@ func initClientInstance(config EtcdGroupServiceConfig, clientOrNil *clientv3.Cli
 	return err
 }
 
-func createBaseClient(config EtcdGroupServiceConfig) (*clientv3.Client, error) {
+func createBaseClient(config config.EtcdGroupServiceConfig) (*clientv3.Client, error) {
 	cli, err := clientv3.New(clientv3.Config{
 		Endpoints:   config.Endpoints,
 		DialTimeout: config.DialTimeout,
