+++ reporters.go
@@ -1,12 +1,14 @@
 package pitaya
 
 import (
+	"github.com/topfreegames/pitaya/v2/config"
 	"github.com/topfreegames/pitaya/v2/logger"
 	"github.com/topfreegames/pitaya/v2/metrics"
+	"github.com/topfreegames/pitaya/v2/metrics/models"
 )
 
 // CreatePrometheusReporter create a Prometheus reporter instance
-func CreatePrometheusReporter(serverType string, config metrics.PrometheusConfig, customSpecs metrics.CustomMetricsSpec) (*metrics.PrometheusReporter, error) {
+func CreatePrometheusReporter(serverType string, config config.PrometheusConfig, customSpecs models.CustomMetricsSpec) (*metrics.PrometheusReporter, error) {
 	logger.Log.Infof("prometheus is enabled, configuring reporter on port %d", config.Port)
 	prometheus, err := metrics.GetPrometheusReporter(serverType, config, customSpecs)
 	if err != nil {
@@ -16,7 +18,7 @@ func CreatePrometheusReporter(serverType string, config metrics.PrometheusConfig
 }
 
 // CreateStatsdReporter create a Statsd reporter instance
-func CreateStatsdReporter(serverType string, config metrics.StatsdConfig) (*metrics.StatsdReporter, error) {
+func CreateStatsdReporter(serverType string, config config.StatsdConfig) (*metrics.StatsdReporter, error) {
 	logger.Log.Infof(
 		"statsd is enabled, configuring the metrics reporter with host: %s",
 		config.Host,
